+++ peritheos/__init__.py
@@ -6,7 +6,7 @@ from peritheos.uncertainty import (
     PredictionUncertainty,
 )
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"
 
 __all__ = [
     "EOSUncertainty",